LPS-156912 Moves logic to display context, no logic changes
